pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sspr*"]

[project]
name = "sspr"
version = "0.0.1"
dependencies = [
    'litellm',
    'loguru',
    'mdformat-gfm',
    'pydantic',
]

[project.optional-dependencies]

testing=[
    "pytest",
]
